requirements.txt
aiosmtpd==1.4.4.post2
azure-identity==1.15.0
ipcalc==1.99.0
logging-handler==1.0.7
parameter-verification==0.1.1
queue-processor==1.1.0
requests==2.31.0
azure-communication-email==1.0.0
azure_smtp_relay>=1.0.0